fastapi_project/requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
Office365-REST-Python-Client
python-dotenv

